Move cuda-mpi to 64-bit
This ensures we can compress files larger than 4gb
note: this commit is still buggy, it seems that the Huffman compression code messes up at the edges so I'll need to fix that

+++ Serial/decompress.c
@@ -10,10 +10,10 @@ struct huffmanTree huffmanTreeNode[512];
 
 int main(int argc, char **argv) {
     clock_t start, end;
-    size_t cpu_time_used;
-    size_t i, j;
+    uint64_t cpu_time_used;
+    uint64_t i, j;
     unsigned int distinctCharacterCount, combinedHuffmanNodes;
-    size_t compressedFileLength, outputFileLength, outputFileLengthCounter,
+    uint64_t compressedFileLength, outputFileLength, outputFileLengthCounter,
         frequency[256] = {0};
     unsigned char currentInputBit, currentInputByte, *compressedData,
         *outputData, bitSequence[255], bitSequenceLength = 0;
@@ -24,13 +24,13 @@ int main(int argc, char **argv) {
     compressedFile = fopen(argv[1], "rb");
 
     // read the header and fill frequency array
-    fread(&outputFileLength, sizeof(size_t), 1, compressedFile);
-    fread(frequency, 256 * sizeof(size_t), 1, compressedFile);
+    fread(&outputFileLength, sizeof(uint64_t), 1, compressedFile);
+    fread(frequency, 256 * sizeof(uint64_t), 1, compressedFile);
 
     // find length of compressed file
     fseek(compressedFile, 0, SEEK_END);
-    compressedFileLength = ftell(compressedFile) - 257 * sizeof(size_t);
-    fseek(compressedFile, 257 * sizeof(size_t), SEEK_SET);
+    compressedFileLength = ftell(compressedFile) - 257 * sizeof(uint64_t);
+    fseek(compressedFile, 257 * sizeof(uint64_t), SEEK_SET);
 
     // allocate required memory and read the file to memoryand then close file
     compressedData = malloc((compressedFileLength) * sizeof(unsigned char));
